Refactor delete fetch
