+++ src/components/TableRow/TableRow.js
@@ -1,25 +1,27 @@
-import React from 'react'
+import { useEffect } from 'react'
 import { useState } from 'react'
 import './TableRow.css'
 import bin from '../../images/bin.png'
 
-const TableRow = ({savedTask, deleteTask, savedTasks, setSavedTasks}) => {
-
-    const handleChange = () => {
-        const allOtherTasks = savedTasks.filter(saved => saved.id !== savedTask.id)
-        const updatedTask = {...savedTask}
-        updatedTask.complete = !updatedTask.complete
-        const updatedTasks = [...allOtherTasks, updatedTask]
-        setSavedTasks(updatedTasks)
-    }
+const TableRow = ({savedTask, deleteTask, handleChange}) => {
     
-  return (
-    <tr className={savedTask.complete ? 'done' : ''} >
+  const change = () => {
+    handleChange(savedTask)
+  }
+
+  const renderRow = () => {
+    return (
+      <tr className={savedTask.complete ? 'done' : ''} >
         <td>{savedTask.category}</td>
         <td>{savedTask.task}</td>
-        <td className='center-clm'><button onClick={handleChange} className={savedTask.complete ? 'check-task checked' : 'check-task' }/></td>
+        <td className='center-clm'><button onClick={change} className={savedTask.complete ? 'check-task checked' : 'check-task' }/></td>
         <td className='center-clm'><button onClick={() => deleteTask(savedTask.id)} className='trash'><img src={bin} className='bin'/></button></td>
     </tr>
+    )
+  }
+
+  return (
+    <>{renderRow()}</>
   )
 }
