+++ src/components/TaskListView/TaskListView.js
@@ -1,8 +1,13 @@
 import './TaskListView.css'
 import TableRow from '../TableRow/TableRow'
 import { deleteSavedTask } from '../apiCalls'
+import { useEffect } from 'react'
 
 const TaskListView = ({savedTasks, setSavedTasks}) => {
+
+  // useEffect(() => {
+
+  // }, [savedTasks])
 
   const deleteTask = (id) => {
     deleteSavedTask(id).then(newSaved => {
@@ -10,8 +15,16 @@ const TaskListView = ({savedTasks, setSavedTasks}) => {
     })
   }
 
-  const rows = savedTasks.map((savedTask, index) => {
-    return (<TableRow savedTask={savedTask} key={`row-${index}`} deleteTask={deleteTask} savedTasks={savedTasks} setSavedTasks={setSavedTasks} />)
+  const handleChange = (savedTask) => {
+    const allOtherTasks = savedTasks.filter(saved => saved.id !== savedTask.id)
+    const updatedTask = {...savedTask}
+    updatedTask.complete = !updatedTask.complete
+    const updatedTasks = [...allOtherTasks, updatedTask]
+    setSavedTasks(updatedTasks)
+  }
+
+  const rows = () => savedTasks.sort((a,b) => b.id - a.id).map((savedTask, index) => {
+    return (<TableRow savedTask={savedTask} key={`row-${index}`} deleteTask={deleteTask} handleChange={handleChange}/>)
   }) 
 
   return (
@@ -24,7 +37,7 @@ const TaskListView = ({savedTasks, setSavedTasks}) => {
         </tr>
       </thead>
       <tbody>
-        {savedTasks.length ? rows : <tr className='no-tasks'>Save a task to view it here!</tr>}
+        {savedTasks.length ? rows() : <tr className='no-tasks'><td>Save a task to view it here!</td></tr>}
       </tbody>
     </table>
   )
